# pos-carts/src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Carts
  description: Carts
  version: 1.0.0
  license:
    name: MIT
paths:
  /carts:
    get:
      summary: Get the current cart
      operationId: getCart
      tags:
        - cart
      responses:
        '200':
          description: A cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /carts/{productId}:
    post:
      summary: Add a product to cart
      operationId: addProductById
      tags:
        - cart
      parameters:
        - name: productId
          in: path
          required: true
          description: The id of the product to add
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Remove a product from cart
      operationId: removeProduct
      tags:
        - cart
      parameters:
        - name: productId
          in: path
          required: true
          description: ID of product to remove
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: double
        image:
          type: string
    Item:
      type: object
      required:
        - product
        - quantity
      properties:
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string